% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsn_mvsnc.R
\name{invmills}
\alias{invmills}
\title{Inverse Mills ratio}
\usage{
invmills(x, mu = 0, sd = 1)
}
\arguments{
\item{x}{Finite numeric scalar.}

\item{mu}{Mean of the normal distribution. Default is \code{0}.}

\item{sd}{Standard deviation of the normal distribution. Default is \code{1}.}
}
\value{
A positive numeric scalar.
}
\description{
Computes the inverse Mills ratio for a normal random variable.
}
\keyword{internal}
