% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mvnc.R
\name{mvn_loglik}
\alias{mvn_loglik}
\title{Matrix-variate normal log-likelihood}
\usage{
mvn_loglik(samples, M, Sigma, Psi)
}
\arguments{
\item{samples}{Numeric array with dimensions \eqn{p \times q \times n}.}

\item{M}{Location matrix of dimension \eqn{p \times q}.}

\item{Sigma}{Row covariance matrix of dimension \eqn{p \times p}.}

\item{Psi}{Column covariance matrix of dimension \eqn{q \times q}.}
}
\value{
Numeric scalar containing the total log-likelihood.
}
\description{
Computes the observed log-likelihood for complete matrix-variate normal data.
}
\keyword{internal}
