% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mvnc.R
\name{solve_sym_pd}
\alias{solve_sym_pd}
\title{Invert a symmetric positive definite matrix}
\usage{
solve_sym_pd(A, epsilon = 1e-08)
}
\arguments{
\item{A}{Numeric square matrix.}

\item{epsilon}{Positive tolerance used for positive-definiteness correction.}
}
\value{
The inverse of \code{A}.
}
\description{
Stabilizes a matrix to be positive definite and computes its inverse using
the Cholesky decomposition.
}
\keyword{internal}
