% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsn_mvsnc.R
\name{sqrtm}
\alias{sqrtm}
\title{Matrix square root}
\usage{
sqrtm(A, epsilon = 1e-08)
}
\arguments{
\item{A}{Numeric scalar or numeric square matrix.}

\item{epsilon}{Numerical tolerance for detecting negative eigenvalues.}
}
\value{
A matrix containing the square root of \code{A}.
}
\description{
Computes the symmetric square root of a nonnegative scalar or symmetric
positive semidefinite matrix.
}
\keyword{internal}
