% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mvnc.R
\name{validate_mvn_input}
\alias{validate_mvn_input}
\title{Validate matrix-variate normal input data}
\usage{
validate_mvn_input(samples)
}
\arguments{
\item{samples}{Numeric array with dimensions \eqn{p \times q \times n}.}
}
\value{
Invisibly returns \code{TRUE} if the input is valid.
}
\description{
Checks whether the sample object is a three-dimensional array without missing
values.
}
\keyword{internal}
