% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mvnc.R
\name{vec_col}
\alias{vec_col}
\title{Vectorize a matrix by columns}
\usage{
vec_col(X)
}
\arguments{
\item{X}{Numeric matrix.}
}
\value{
A numeric vector containing the entries of \code{X} stacked by columns.
}
\description{
Converts a matrix into a vector using column-wise ordering.
}
\keyword{internal}
