% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrsn.R
\name{loglik_mvrsn}
\alias{loglik_mvrsn}
\title{Log-likelihood function of the MVRSN distribution}
\usage{
loglik_mvrsn(X_array, M, A, Sigma, Psi)
}
\arguments{
\item{X_array}{Sample array with dimensions p x q x n.}

\item{M}{Location matrix.}

\item{A}{Skewness matrix.}

\item{Sigma}{Positive definite row covariance matrix.}

\item{Psi}{Positive definite column covariance matrix.}
}
\value{
Observed-data log-likelihood.
}
\description{
Sums the closed-form log-density over a sample array p x q x n.
}
