% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_mvnc.R
\name{mvn_ecm}
\alias{mvn_ecm}
\title{ECM estimation for the matrix-variate normal model}
\usage{
mvn_ecm(samples, precision = 1e-06, MaxIter = 50)
}
\arguments{
\item{samples}{Numeric array with dimensions \eqn{p \times q \times n}.}

\item{precision}{Positive scalar. Convergence tolerance.}

\item{MaxIter}{Positive integer. Maximum number of ECM iterations.}
}
\value{
An object of class \code{"MVN.ECM"} containing:
\describe{
\item{mu}{Estimated location matrix.}
\item{Sigma}{Estimated row covariance matrix.}
\item{Psi}{Estimated column covariance matrix.}
\item{loglik}{Final log-likelihood.}
\item{BIC}{Bayesian information criterion.}
\item{iter}{Number of iterations performed.}
\item{converged}{Logical value indicating whether convergence was reached.}
}
}
\description{
Fits a complete-data matrix-variate normal model by an ECM algorithm.
}
\keyword{internal}
