% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrsn.R
\name{mvrsn_mean}
\alias{mvrsn_mean}
\title{Mean matrix of the MVRSN distribution}
\usage{
mvrsn_mean(M, A)
}
\arguments{
\item{M}{Location matrix.}

\item{A}{Skewness matrix.}
}
\value{
The theoretical mean matrix.
}
\description{
For standard half-normal latent variables, E(W_i) = sqrt(2 / pi). Therefore
E(X) = M + sqrt(2 / pi) A.
}
