% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsn.R
\name{rmvsn}
\alias{rmvsn}
\title{Generate Matrix Variate Skew Normal Samples (Complete Data)}
\usage{
rmvsn(n = n, M = M, A = A, U = U, V = V)
}
\arguments{
\item{n}{Integer. Number of matrices to generate.}

\item{M}{Mean matrix of dimension \eqn{p \times q}.}

\item{A}{Skewness matrix of dimension \eqn{p \times q}.}

\item{U}{Row covariance matrix \eqn{p \times p}.}

\item{V}{Column covariance matrix \eqn{q \times q}.}
}
\value{
A 3D array of generated samples with dimensions
\eqn{p \times q \times n}.
}
\description{
Generates a sample of matrices following the matrix-variate skew-normal
model \eqn{X = M + A|W| + \mathcal{V}}, where \eqn{W \sim N(0,1)} and
\eqn{\mathcal{V} \sim \mathcal{MVN}_{p \times q}(0, U, V)}.
}
\details{
To be more precise, this function simulates a sample of size \eqn{n}
using the stochastic representation:
\deqn{X_{i} = M + A|W_{i}| + \mathcal{V}_{i}}
where \deqn{\mathcal{V}_{i}} is generated using \code{rmatrixnorm()},
which is provided by the \pkg{LaplacesDemon} package.
}
\examples{
p <- 3
q <- 2
n <- 20
M <- matrix(0, p, q)
A <- matrix(1, p, q)
U <- diag(p)
V <- diag(q)

samples <- rmvsn(n = n, M = M, A = A, U = U, V = V)
dim(samples)   # Should be 3 x 2 x 20
}
