% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_generator.R
\name{rmvst}
\alias{rmvst}
\title{Generate matrix-variate skew-t samples}
\usage{
rmvst(n, M, A, U, V, nu)
}
\arguments{
\item{n}{Positive integer. Number of matrices to generate.}

\item{M}{Numeric matrix of dimension \eqn{p \times q}. Location matrix.}

\item{A}{Numeric matrix of dimension \eqn{p \times q}. Skewness matrix.}

\item{U}{Square numeric matrix of dimension \eqn{p \times p}. Row covariance matrix.}

\item{V}{Square numeric matrix of dimension \eqn{q \times q}. Column covariance matrix.}

\item{nu}{Positive numeric scalar. Degrees-of-freedom parameter.}
}
\value{
A numeric array of dimension \eqn{p \times q \times n}.
}
\description{
Generates random matrices from a matrix-variate skew-t-type distribution using
the stochastic representation
}
\details{
\deqn{
X = M + \frac{A |Z| + E}{\sqrt{W}},
}

where \eqn{Z \sim N(0,1)},
\eqn{E \sim MN_{p \times q}(0, U, V)}, and
\eqn{W \sim Gamma(\nu/2, \nu/2)}.
}
\keyword{internal}
